Create highlight sectors
