+++ CadRevealComposer/Operations/StidTagMapper/TagDataFromStid.cs
@@ -5,29 +5,29 @@ using Newtonsoft.Json;
 [JsonObject]
 public class TagDataFromStid
 {
-    public string TagNo { get; set; }
-    public string Description { get; set; }
-    public string TagStatus { get; set; }
-    public int TagCategory { get; set; }
-    public string TagCategoryDescription { get; set; }
-    public string TagType { get; set; }
-    public string UpdatedDate { get; set; }
-    public string LocationCode { get; set; }
-    public string DisciplineCode { get; set; }
-    public string ContrCode { get; set; }
-    public string System { get; set; }
-    public string ProjectCode { get; set; }
-    public string PoNo { get; set; }
-    public string PlantNo { get; set; }
+    public required string TagNo { get; set; }
+    public required string Description { get; set; }
+    public required string TagStatus { get; set; }
+    public required int TagCategory { get; set; }
+    public required string TagCategoryDescription { get; set; }
+    public required string? TagType { get; set; }
+    public required string UpdatedDate { get; set; }
+    public required string LocationCode { get; set; }
+    public required string DisciplineCode { get; set; }
+    public required string ContrCode { get; set; }
+    public required string System { get; set; }
+    public required string ProjectCode { get; set; }
+    public required string PoNo { get; set; }
+    public required string PlantNo { get; set; }
     public float? XCoordinate { get; set; }
     public float? YCoordinate { get; set; }
     public float? ZCoordinate { get; set; }
-    public AdditionalFields[] AdditionalFields { get; set; }
+    public required AdditionalFields[] AdditionalFields { get; set; }
 }
 
 [JsonObject]
 public class AdditionalFields
 {
-    public string Type { get; set; }
-    public string Value { get; set; }
+    public required string Type { get; set; }
+    public required string Value { get; set; }
 }
