+++ CadRevealComposer/SceneCreator.cs
@@ -138,7 +138,7 @@ public static class SceneCreator
         JsonUtils.JsonSerializeToFile(scene, scenePath, writeIndented: EnvUtil.IsDebugBuild); // We don't want indentation in prod, it doubles the size. Format in an editor if needed.
     }
 
-    public static void ExportSectorGeometries(
+    private static void ExportSectorGeometries(
         IReadOnlyList<APrimitive> geometries,
         string sectorFilename,
         string? outputDirectory
